fix(test): split branded asset contract coverage
